Move to NPM from yarn
Since we have to release shopkeeper in order for it to be used,
it's better to use npm for managing this project. Yarn doesn't play
nicely with the GitHub Package Registry and all the documentation is for
NPM so rather than try to be special, we'll just do what is well
supported.