Pass the environment on download
We regularly see build failures due to a theme with the name not being
found. It's a flakey failure so we it's been hard to track down.

Rather than depend on the live flag being passed to Themekit, we'll pass
the environment ourselves. Let's see if this prevents the regular
failures.