Restrict deployment strategies
We want to provide better error messages when an unsupported
strategy is passed. By putting this restriction at the flag parser
level, we don't need to add a guard to ensure the strategy is one of the
expected ones. Also, this prevents people from getting an unexpected
result because our switch statement in the implementation that defaults
to the basic strategy. This could have an unexpected result.