Update theme init to affect .gitignore
When we install Shopkeeper, we want our .gitignore to match the
convention of storing settings in .shopkeeper.

This is a first pass of this feature. Writing code on behalf of a user
is always a little tricky. At this point, we're doing a naive append or
create. I'm open to more sophistication in the future, but for a command
that's run infrequently a project, it's not worth the effort right now.