Set Shopkeeper to depend on the latest Shopify 3.x release
Previously, we limited the versions to patch updates of 3.48. The
reality is Shopify continues to interate at a quick pace and I'm not
sure we want to constrain ourselves to the minor version. This would
require a fresh Shopkeeper release each time Shopify increments the
minor version. This would add extra maintenance overhead. If Shopify
gets carless with it's releases we'll lock things down. However, being
the pros they are, I don't think this is necessary.